requirements.txt
Django==4.0.4
django-filter==21.1
django-mptt==0.13.4
djangorestframework==3.13.1
